refactor(crm): migrate canAccessDeal utility to crmUtil and remove dealAccessUtil
