--- frontend/src/community/crm/utils/dealAccessUtil.ts
@@ -1,11 +1,0 @@
-interface DealAccessSession {
-  userId?: number;
-  isCrmAdmin?: boolean;
-  isCrmSalesManager?: boolean;
-}
-
-export const canAccessDeal = (
-  ownerEmployeeId: number,
-  { userId, isCrmAdmin, isCrmSalesManager }: DealAccessSession
-): boolean =>
-  Boolean(isCrmAdmin || isCrmSalesManager) || ownerEmployeeId === userId;
